+++ src/components/app/App.js
@@ -1,4 +1,4 @@
-import Board from "../Board";
+import Board from "./board/Board";
 
 import React from "react";
