+++ src/components/app/board/_board.scss
