+++ src/components/app/board/list/_card.scss
