+++ src/components/app/board/list/_list.scss
